fixed conda-env-activate-for-buffer behavior
Once a conda environment is enabled for a buffer this code forcibly
set the conda-project-env-path buffer-local variable.

This causes conda-env-activate-for-buffer to not be able to activate
a different environment from a YAML file on the path.

The conda-project-env-path variable should be only set by a definition
from a.dir-locals.el file, which should indeed be able to override the
environment defined in an environment.yaml file.