Register process sentinel to log status on process exit
Enables lexical scoping, which is needed for closures to work in sentinel callbacks.

The complexity in this handler is due to the need to ensure that any sentinels
already registered for the process are correctly preserved.

For example, I personally have a hook for term-mode, which registers a sentinel
to prompt for closing the buffer on process exit. Without this logic the logger
for the completion status would override my handler completely.